requirements.txt
librosa
pandas
numpy
scikit-learn
keras
ipython
matplotlib
seaborn
tensorflow
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
plotly